requirements.txt
pytesseract
pillow
pdf2image
opencv-python
rich

